inst_macros::errors

Enum ImplBug

pub(crate) enum ImplBug {

    FileContainsNoImpl {},
    DelegateMacroNotInitialized {},
    DelegateMacroNotFound {},
    InstanceImplMacroNotFound {},
    InstanceImplNodeMacroNotFound {},
    InstanceImplNodeMacroArgsFound {},
    InstanceImplMacroArgsNotFound {},
    InstanceImplMacroIndexesParseFailed {},
    InstanceLastImplMacroNotFound {},
    InstanceLastImplMacroArgNotFound {},
    InstanceLastImplMacroIndexParseFailed {},
    InstanceLastImplDelegateMacroNotFound {},
    InstanceLastImplMacroIndexInconsistent {},
    InstanceImplIndexesArgsLenInconsistent {},
    InstanceImplIndexesArgsWrong {},
    TraitImplExpected {},
    TraitImplAngleArgsExpected {},
    NodeAttrNotRemoved {},
    TupleAttrNotRemoved {},
    TransformedFileInconsistent {},
    TransformedFileShouldNotHaveDocTarget {},
    TransformedFileDoesNotHaveDocTarget {},
    TransformedFileDoesNotHaveNotDocTarget {},
    NotDocTargetDoesNotHaveDocs {},
    InstanceImplCounterGenericArgMissing {},
    InstanceImplCounterGenericArgNotConst {},
    InstanceImplCounterGenericArgNotLit {},
    InstanceImplCounterGenericArgNotLitInt {},
    InstanceImplCounterGenericArgInvalid {},
    InstanceImplCounterIdentConstItemMissing {},
    InstanceImplCounterIdentConstNotLit {},
    InstanceImplCounterIdentConstNotLitBStr {},
    InstanceImplCounterIdentConstInvalid {},
    InstanceImplCounterIdentConstTypeInvalid {},
    InstanceImplCounterIdentConstCounterAttrMissing {},
    InstanceImplCounterIdentConstCounterAttrInvalid {},
    InstanceImplCounterIdentConstCounterAttrArgNotParsed {},
    InstanceImplCounterIdentConstCounterAttrArgInvalid {},
    InstanceImplInstanceCounterTypeMissing {},
    InstanceImplInstanceCounterTypeInvalid {},
    InconsistencyFindingNodeAttribute {},
    ExpectedNodeTypeImpl {},
    ExpectedNodeTypeImplSub {},
    ExpectedNodeTypeImplPub {},
    ExpectedNodeTypeImplPubAttrList {},
    ExpectedNodeTypeImplSubAttrPath {},
    EitherTupleOrNode {},
    InconsistencyFindingTupleAttribute {},
    ExpectedTupleTypeImpl {},
    TupleSubAssocQualifiedButNotTrait {},
    DocFileTooManyItemsToTransform {},
    DocFileContainsNoImpl {},
    DocImplExpectedToContainDocAttr {},
    DocFileExpectedToBeEmpty {},
    ImplTupleAddonNotInstFeatureGated {},
    ImplTupleAddonNonCamelCaseAttrMissing {},
    TupleEnumAddonInstImplsInconsistent {},
    TupleEnumAddonInstImplItemsExtraFound {},
    ImplTupleAddonInstImplNotInstAttributed {},
    ImplTupleAddonInstImplInstArgInconsistent {},
    ImplTupleAddonInstImplNotTraitImpl {},
    ImplTupleAddonInstTraitImplInvalidTrait {},
    TupleEnumAddonInstImplItemNotType {},
    TupleEnumAddonInstImplItemTypeInvalid {},
    TupleEnumAddonInstImplItemIdentInvalid {},
    TupleEnumAddonVariantIndexInvalid {},
    TupleEnumAddonVariantIdentInvalid {},
    TupleEnumAddonVariantFieldsNotUnamed {},
    TupleEnumAddonVariantFieldInconsistent {},
    TupleEnumAddonVariantTypeInvalid {},
    TupleEnumInstNodeTypeNotFound {},
    TupleEnumInstNodeTypePubNodeAttrsNotFound {},
    TupleEnumAddonVariantLastInstPunctNotTerminated {},
    TupleEnumAddonVariantInstPunctInvalid {},

}

Variants

FileContainsNoImpl

DelegateMacroNotInitialized

DelegateMacroNotFound

InstanceImplMacroNotFound

InstanceImplNodeMacroNotFound

InstanceImplNodeMacroArgsFound

InstanceImplMacroArgsNotFound

InstanceImplMacroIndexesParseFailed

InstanceLastImplMacroNotFound

InstanceLastImplMacroArgNotFound

InstanceLastImplMacroIndexParseFailed

InstanceLastImplDelegateMacroNotFound

InstanceLastImplMacroIndexInconsistent

InstanceImplIndexesArgsLenInconsistent

InstanceImplIndexesArgsWrong

TraitImplExpected

TraitImplAngleArgsExpected

NodeAttrNotRemoved

TupleAttrNotRemoved

TransformedFileInconsistent

TransformedFileShouldNotHaveDocTarget

TransformedFileDoesNotHaveDocTarget

TransformedFileDoesNotHaveNotDocTarget

NotDocTargetDoesNotHaveDocs

InstanceImplCounterGenericArgMissing

InstanceImplCounterGenericArgNotConst

InstanceImplCounterGenericArgNotLit

InstanceImplCounterGenericArgNotLitInt

InstanceImplCounterGenericArgInvalid

InstanceImplCounterIdentConstItemMissing

InstanceImplCounterIdentConstNotLit

InstanceImplCounterIdentConstNotLitBStr

InstanceImplCounterIdentConstInvalid

InstanceImplCounterIdentConstTypeInvalid

InstanceImplCounterIdentConstCounterAttrMissing

InstanceImplCounterIdentConstCounterAttrInvalid

InstanceImplCounterIdentConstCounterAttrArgNotParsed

InstanceImplCounterIdentConstCounterAttrArgInvalid

InstanceImplInstanceCounterTypeMissing

InstanceImplInstanceCounterTypeInvalid

InconsistencyFindingNodeAttribute

ExpectedNodeTypeImpl

ExpectedNodeTypeImplSub

ExpectedNodeTypeImplPub

ExpectedNodeTypeImplPubAttrList

ExpectedNodeTypeImplSubAttrPath

EitherTupleOrNode

InconsistencyFindingTupleAttribute

ExpectedTupleTypeImpl

TupleSubAssocQualifiedButNotTrait

DocFileTooManyItemsToTransform

DocFileContainsNoImpl

DocImplExpectedToContainDocAttr

DocFileExpectedToBeEmpty

ImplTupleAddonNotInstFeatureGated

ImplTupleAddonNonCamelCaseAttrMissing

TupleEnumAddonInstImplsInconsistent

TupleEnumAddonInstImplItemsExtraFound

ImplTupleAddonInstImplNotInstAttributed

ImplTupleAddonInstImplInstArgInconsistent

ImplTupleAddonInstImplNotTraitImpl

ImplTupleAddonInstTraitImplInvalidTrait

TupleEnumAddonInstImplItemNotType

TupleEnumAddonInstImplItemTypeInvalid

TupleEnumAddonInstImplItemIdentInvalid

TupleEnumAddonVariantIndexInvalid

TupleEnumAddonVariantIdentInvalid

TupleEnumAddonVariantFieldsNotUnamed

TupleEnumAddonVariantFieldInconsistent

TupleEnumAddonVariantTypeInvalid

TupleEnumInstNodeTypeNotFound

TupleEnumInstNodeTypePubNodeAttrsNotFound

TupleEnumAddonVariantLastInstPunctNotTerminated

TupleEnumAddonVariantInstPunctInvalid

Implementations

impl ImplBug

pub fn to_diagnostic(&self) -> Diagnostic

Converts the enum variant into a [Diagnostic] with all configured message, tags, span, help, and notes.

pub fn to_string(&self) -> String

Renders the diagnostic as a plain string.

Spanned helps are omitted since they rely on compiler diagnostics and are not meaningful in runtime panic messages.