inst_macros::errors

Enum ImplError

pub(crate) enum ImplError {

    NodeRequiresArgs {
        attr: Attribute,
    },
    NodeRequiresArgsNotValue {
        meta: MetaNameValue,
    },
    ExpectedNodeTypeImpl {
        ty: Type,
    },
    ExpectedTupleTypeImpl {
        ty: Type,
    },
    TupleArgsAreInfered {
        attr: Attribute,
    },
    TupleArgMustBeStruct {
        ty: Type,
    },
    TupleArgMustNotBeQualified {
        qself: Type,
    },
    TupleArgPathQualifyElsewhere {
        seg: PathSegment,
    },
    TupleArgGenericArgsNotSupported {
        path_args: PathArguments,
    },
    QualifiedButNotTrait {
        qself: Type,
    },
    TupleSubAssocQualifiedButNotTrait {
        qself: Type,
    },
    TupleSubAssocMustBeSelfQualified {
        qself: Type,
    },
    GenericFoundExpectsTraitQualifier {
        seg: PathSegment,
    },
    InherentImplExpected {
        trait_: Path,
    },
    TraitImplExpected {
        self_ty: Type,
    },
    NoParenTraitImpl {
        seg: PathSegment,
    },
    NodeOnlyInTypeAssoc {
        attr: Attribute,
    },
    TupleOnlyInTypeAssoc {
        attr: Attribute,
    },
    InstEitherTraitOrNodeOrTuple {},
    DoubleNode {
        attr: Attribute,
    },
    EitherTupleOrNode {
        attr: Attribute,
    },
    WhereClauseNotSupportInTuple {
        where_: WhereClause,
    },
    DoubleTuple {
        attr: Attribute,
    },

}

Variants

NodeRequiresArgs

Fields

attr: Attribute

NodeRequiresArgsNotValue

Fields

meta: MetaNameValue

ExpectedNodeTypeImpl

Fields

ty: Type

ExpectedTupleTypeImpl

Fields

ty: Type

TupleArgsAreInfered

Fields

attr: Attribute

TupleArgMustBeStruct

Fields

ty: Type

TupleArgMustNotBeQualified

Fields

qself: Type

TupleArgPathQualifyElsewhere

Fields

seg: PathSegment

TupleArgGenericArgsNotSupported

Fields

path_args: PathArguments

QualifiedButNotTrait

Fields

qself: Type

TupleSubAssocQualifiedButNotTrait

Fields

qself: Type

TupleSubAssocMustBeSelfQualified

Fields

qself: Type

GenericFoundExpectsTraitQualifier

Fields

seg: PathSegment

InherentImplExpected

Fields

trait_: Path

TraitImplExpected

Fields

self_ty: Type

NoParenTraitImpl

Fields

seg: PathSegment

NodeOnlyInTypeAssoc

Fields

attr: Attribute

TupleOnlyInTypeAssoc

Fields

attr: Attribute

InstEitherTraitOrNodeOrTuple

DoubleNode

Fields

attr: Attribute

EitherTupleOrNode

Fields

attr: Attribute

WhereClauseNotSupportInTuple

Fields

where_: WhereClause

DoubleTuple

Fields

attr: Attribute

Implementations

impl ImplError

pub fn to_diagnostic(&self) -> Diagnostic

Converts the enum variant into a [Diagnostic] with all configured message, tags, span, help, and notes.

pub fn to_string(&self) -> String

Renders the diagnostic as a plain string.

Spanned helps are omitted since they rely on compiler diagnostics and are not meaningful in runtime panic messages.