inst_macros::errors

Enum TraitError

pub(crate) enum TraitError {

    NodeRequiresArgs {
        attr: Attribute,
    },
    NodeRequiresArgsNotValue {
        meta: MetaNameValue,
    },
    TupleRequiresArgs {
        attr: Attribute,
    },
    TupleRequiresArgsNotValue {
        meta: MetaNameValue,
    },
    Max4Params {
        idents: IdentList,
    },
    InstTraitOverflowReference {
        ident: Ident,
    },
    SumNoArgs {
        attr: Attribute,
    },
    SumNotAllowed {
        attr: Attribute,
    },
    DoubleNode {
        attr: Attribute,
    },
    DoubleTuple {
        attr: Attribute,
    },
    NodeOnlyInTypeAssoc {
        attr: Attribute,
    },
    TupleOnlyInTypeAssoc {
        attr: Attribute,
    },
    InstEitherTraitOrNodeOrTuple {},
    DuplicateInstanceTraitBound {
        bound: TraitBound,
    },
    ExpectedInstTraitNodeBound {
        bounds: Punctuated<TypeParamBound, Plus>,
        exp: Punctuated<Ident, PathSep>,
    },
    ParenthesizedInstTraitNodeBound {
        paren: ParenthesizedGenericArguments,
    },
    EitherTupleOrNode {
        attr: Attribute,
    },
    WhereClauseNotSupportInTuple {
        where_: WhereClause,
    },

}

Variants

NodeRequiresArgs

Fields

attr: Attribute

NodeRequiresArgsNotValue

Fields

meta: MetaNameValue

TupleRequiresArgs

Fields

attr: Attribute

TupleRequiresArgsNotValue

Fields

meta: MetaNameValue

Max4Params

Fields

idents: IdentList

InstTraitOverflowReference

Fields

ident: Ident

SumNoArgs

Fields

attr: Attribute

SumNotAllowed

Fields

attr: Attribute

DoubleNode

Fields

attr: Attribute

DoubleTuple

Fields

attr: Attribute

NodeOnlyInTypeAssoc

Fields

attr: Attribute

TupleOnlyInTypeAssoc

Fields

attr: Attribute

InstEitherTraitOrNodeOrTuple

DuplicateInstanceTraitBound

Fields

bound: TraitBound

ExpectedInstTraitNodeBound

Fields

bounds: Punctuated<TypeParamBound, Plus>

exp: Punctuated<Ident, PathSep>

ParenthesizedInstTraitNodeBound

Fields

paren: ParenthesizedGenericArguments

EitherTupleOrNode

Fields

attr: Attribute

WhereClauseNotSupportInTuple

Fields

where_: WhereClause

Implementations

impl TraitError

pub fn to_diagnostic(&self) -> Diagnostic

Converts the enum variant into a [Diagnostic] with all configured message, tags, span, help, and notes.

pub fn to_string(&self) -> String

Renders the diagnostic as a plain string.

Spanned helps are omitted since they rely on compiler diagnostics and are not meaningful in runtime panic messages.